Remove ManagedThread's direct access to Thread internals
