+++ src/main/bindings/rust/wrapper.rs
@@ -1676,6 +1676,9 @@ extern "C" {
     pub fn thread_getSysCallCondition(thread: *mut Thread) -> *mut SysCallCondition;
 }
 extern "C" {
+    pub fn thread_clearSysCallCondition(thread: *mut Thread);
+}
+extern "C" {
     pub fn thread_getSignalSet(thread: *mut Thread) -> *mut sigset_t;
 }
 extern "C" {
